Database.Orville.PostgreSQL.Conduit

Description

Documentation

selectConduit :: (Monad m, MonadOrville conn m, MonadCatch m, MonadResource m) => Select row -> ConduitT () row m ()

selectConduit provides a way to stream the results of a Select query from the database one by one using the conduit library. You can fuse the conduit built by this function with your own conduit pipeline to handle rows individually in whatever fashion you need (e.g. turning them into rows of CSV). This is useful if you want to be able to process many rows one by one. You can aggregate the results however you require as part of the conduit processing and then use runConduit (or runConduitRes) from the conduit library to execute the processing pipeline. Alternatively, your web server (wai, servant, etc) may provide support for converting a conduit into a streaming HTTP response.

Beware: this function must load all the results into memory before streaming can begin. For why, see https://www.postgresql.org/docs/9.2/libpq-single-row-mode.html. If memory use is a concern, try streamPages instead.

streamPages

Arguments

:: (MonadOrville conn m, Bounded orderField, Enum orderField)

=> TableDefinition readEnt write key

-> FieldDefinition NotNull orderField

-> (readEnt -> orderField)

-> Maybe WhereCondition

-> Word
number of rows fetched per page

-> ConduitT () readEnt m ()

Build a conduit source that is fed by querying one page worth of results at a time. When the last row of the last page is consumed, the stream ends.