Database.Orville.PostgreSQL.Connection

Description

Documentation

createConnectionPool

Arguments

:: Int
Number of stripes in the connection pool

-> NominalDiffTime
Linger time before closing an idle connection

-> Int
Max number of connections to allocate per stripe

-> String
A PostgreSQL connection string

-> IO (Pool Connection)

createConnectionPool allocates a pool of connections to a PosgreSQL server. The returned pool can be used as the endpoint to newOrvilleEnv to construct.

data Pool a

data Connection