Database.Orville.PostgreSQL.Core

Description

Migration Guide: Although not all exports are identical, most of the items in this module can now be imported from Orville.PostgreSQL.

Please note that the new LibPQ-based version of orville represents a complete re-write of Orville from the ground up. As such many of the APIs have been re-thought with the goal of providing stability and better experience long term.

Major changes:

The library no longer allows the connection type to vary. It is intended only to be used with PostgreSQL. Thus MonadOrville and other types that used to have a conn type parameter no longer have that parameter.

OrvilleT has been removed in favor of simply using ReaderT. For trivial cases (i.e. ReaderT over IO) a pre-packaged Orville monad is provided.

In TableDefinition, the order of the type parameters has changed from TableDefinition readEnity writeEntity key to TableDefinition key writeEntity readEntity. This make it more consistent with the order of these arguments in other types. TableParams has been removed in favor of building a basic table definition with the require parameters first and adding or setting other optional items after initial construction.

RelationalMap has been replaced by SqlMarshaller. Many functions have been renamed, but most functions have a direct or nearly direct translation from the old ones. See the docs on the individual functions such as attrField to see what has changed.

The auto-migration system is significantly improved. Standard indexes no longer need to be given explicit names. Indexes and constraints are now attached to TableDefinition rather than being their own schema items. Indexes and constraints are no longer dropped explicitly -- they are dropped automatically by Orville when they have been removed from the table definiton.

A number of places that previously accepted [] now require NonEmpty instead. This is done in places where an empty list would not be valid SQL. For examples of this change see insertRecordMany and updateFields.

whereAnd and whereOr (which took lists) have been replaced with binary boolean logic functions andExpr and orExpr. These functions also have operator aliases ((.&&), (.||)).

The following items exported from this module have migration guide notes available in their documentation:

TableDefinition

mkTableDefinition

TableParams

RelationalMap

fields

mapAttr

mapField

attrField

maybeMapper

prefixMap

partialMap

readOnlyMap

readOnlyField

OrvilleEnv

newOrvilleEnv

setStartTransactionSQL

aroundRunningQuery

addTransactionCallBack

OrvilleT

HasOrvilleContext

MonadOrville

runOrville

mapOrvilleT

MonadOrvilleControl

defaultLiftWithConnection

defaultLiftFinally

withCachedConnection

withTransaction

ColumnFlag

FieldDefinition

isFieldNullable

fieldOfType

textField

fixedTextField

unboundedTextField

dayField

utcTimeField

int32Field

int64Field

doubleField

boolField

automaticIdField

searchVectorField

nullableField

foreignKeyField

withFlag

withName

withConversion

fieldFromSql

fieldToSqlValue

SomeField

withPrefix

fieldFlags

uniqueIndex

simpleIndex

uniqueConstraint

dropConstraint

SchemaItem

SchemaDefinition

Record

WhereCondition

whereAnd

whereOr

whereIn

whereLike

whereLikeInsensitive

whereNotIn

whereQualified

whereRaw

whereToSql

isNull

isNotNull

migrateSchema

generateMigrationPlan

MigrationPlan

MigrationItem

migrationPlanItems

selectAll

selectFirst

deleteRecord

deleteWhere

findRecord

findRecords

findRecordsBy

insertRecord

insertRecordMany

insertRecordManyReturning

updateFields

updateRecord

createIndexesConcurrently

dropIndexesConcurrently

Documentation

data TableDefinition readEntity writeEntity key

Migration Guide: TableDefinition can now be imported from PostgreSQL. The order of the type parameters has changed from TableDefinition readEnity writeEntity key to TableDefinition key writeEntity readEntity. In the new Orville tables without primary keys are supported, so the key parameter must now be instantiated as either HasKey keyType or NoKey.

A TableDefinition is the center of the Orville universe. A TableDefinition defines the structure of a table in the database and associates it with a Haskell datatype, usually a Haskell record type. The TableDefinition must specify how the Haskell type is converted to and from the database schema, as as well as provide same basic utility functions required by Orville for interacting with the Haskell datatype.

Usually you will use TableParams to construct a TableDefinition in a more concise way. This type is provided as an escape hatch for any situations where TableParams is too restrictive for the sql mapping required by a type.

Constructors

TableDefinition

Fields

tableName :: String

The name of the table in the database.

tableFields :: [SomeField]

A list of field definitions defining the table structure

tableSafeToDelete :: [String]

A list of any columns that may be deleted from the table by Orville. (Orville will never delete a column without being told it is safe)

tablePrimaryKey :: PrimaryKey key

The statically typed field definition that is the primary key. Currently this field must still by listed in tableFields

tableFromSql :: FromSql readEntity

A definition of how to convert the haskell type from a sql row

tableToSql :: ToSql writeEntity ()

A function to set the key on the entity

tableGetKey :: readEntity -> key

A function to get the key on the entity

tableComments :: TableComments ()

Any comments that might be interesting for developers to see. These comments will get printed in the log if there is an erro while attempting to migrate the table.

data PrimaryKey key

primaryKeyIn :: PrimaryKey key -> [key] -> WhereCondition

primaryKeyIn builds a WhereCondition that will match all rows where the primary key is equal to one of the given values. For single-field primary keys this is equivalent to whereIn, but primaryKeyIn also handles composite primary keys.

primaryKeyEquals :: PrimaryKey key -> key -> WhereCondition

primaryKeyEquals builds a WhereCondition that will match the row where the primary key is equal to the given value. For single-field primary keys this is equivalent to .==, but 'primaryKeyEquals also handles composite primary keys.

primaryKeyDescription :: PrimaryKey key -> String

primaryKeyDescription builds a user-readable representation of the primary key for use in error messages and such. It is a comma-delimited list of the names of the fields that make up the primary key.

primaryKeyToSql :: PrimaryKey key -> key -> [SqlValue]

primaryKeyToSql converts a Haskell value for a primary key into the (possibly multiple) sql values that represent the primary key in the database.

primaryKey :: FieldDefinition NotNull key -> PrimaryKey key

primaryKey constructs a single-field primary key from the FieldDefinition that corresponds to the primary key's column. This is generally used while building a TableDefinition.

compositePrimaryKey :: PrimaryKeyPart key -> [PrimaryKeyPart key] -> PrimaryKey key

compositePrimaryKey constructs a multi-field primary key from the given parts, each of which corresponds to one field in the primary key. You should use this while building a TableDefinition for a table that you want to have a multi-column primary key. See primaryKeyPart for how to build the parts to be passed as parameters. Note: there is no special significance to the first argument other than requiring that there is at least one field in the primary key.

primaryKeyPart :: (key -> part) -> FieldDefinition NotNull part -> PrimaryKeyPart key

primaryKeyPart builds on section of a composite primary key based on the field definition that corresponds to that column of the primary key. The function given is used to decompose the Haskell value for the composite key into the individual parts so they can be converted to sql for things like building WhereCondition

mkTableDefinition :: TableParams readEntity writeEntity key -> TableDefinition readEntity writeEntity key

Migration Guide: This function has in the new orville to take the table name, primary key definition and a SqlMarshaller (formerly RelationalMap). Other options such as constraints, indexes, and columns to drop can be added to the TableDefinition after the initial instantiation. The TableParams type has been dropped for the new orville.

mkTableDefinition converts a TableParams to TableDefinition. Usually this is used directly on a record literal of the TableParams. For example:

 data Foo key = Foo key { fooId :: Record }
 myTable :: TableDefinition Foo
 myTable = mkTableDefinition $
   TableParams
     { tblName = "foo"
     , tblMapper = User $ attrField fooId idField
     , tableSafeToDelete = []
     , tblSetKey = key foo -> foo { fooId = key }
     , tblGetKey = fooId
     , tblComments = []
     }

data SqlType a

SqlType defines the mapping of a Haskell type (a) to a SQL column type in the database. This includes both how to convert the type to and from the raw values read from the database as well as the schema information required to create and migrate columns using the type.

Constructors

SqlType

Fields

sqlTypeDDL :: String

The raw SQL DDL to use when creating/migrating columns of this type (not including any NULL or NOT NULL declarations)

sqlTypeReferenceDDL :: Maybe String

The raw SQL DDL to use when creating/migrating columns with foreign keys to this type. This is used foreignRefType to build a new SqlType when making foreign key fields

sqlTypeId :: SqlTypeId

sqlTypeId will be compared to the colType field found in the SqlColDesc return by describeTable when determining whether a column type change is required when migrating the database.

sqlTypeSqlSize :: Maybe Int

'sqlTypeSqlSize will be compared to the colSize field found in the SqlColDesc return by describeTable when determining whether a column type change is required when migrating the database.

sqlTypeToSql :: a -> SqlValue

A function for converting Haskell values of this type into values to be stored in the database.

sqlTypeFromSql :: SqlValue -> Either RowDataErrorReason a

A function for converting values of this are stored in the database into Haskell values. This function should return 'Left RowDataErrorReason' to indicate an error if the conversion is impossible. Otherwise it should return Right the corresponding a value.

serial :: SqlType Int32

serial defines a 32-bit auto-incrementing column type. This corresponds to the SERIAL type in PostgreSQL.

bigserial :: SqlType Int64

bigserial defines a 64-bit auto-incrementing column type. This corresponds to the BIGSERIAL type in PostgresSQL.

text :: Int -> SqlType Text

text defines a fixed length text field type. This corresponds to a "CHAR(len)" type in SQL.

varText :: Int -> SqlType Text

varText defines a variable text field type with a max length. This corresponds to a "VARCHAR(len)" type in SQL.

unboundedText :: SqlType Text

unboundedText defines a fixed length text field type. This corresponds to a TEXT type in PostgreSQL.

integer :: SqlType Int32

integer defines a 32-bit integer type. This corresponds to the INTEGER type in SQL.

bigInteger :: SqlType Int64

bigInteger defines a 64-bit integer type. This corresponds to the BIGINT type in SQL.

double :: SqlType Double

double defines a floating point numeric type. This corresponds to the "DOUBLE PRECISION" type in SQL.

boolean :: SqlType Bool

boolean defines a True/False boolean type. This corresponds to the BOOLEAN type in SQL.

date :: SqlType Day

date defines a type representing a calendar date (without time zone). It corresponds to the DATE type in SQL.

timestamp :: SqlType UTCTime

timestamp defines a type representing a particular point in time (without time zone). It corresponds to the "TIMESTAMP with time zone" type in SQL.

Note: This is NOT a typo. The "TIMESTAMP with time zone" type in SQL does not include any actual time zone information. For an excellent explanation of the complexities involving this type, please see Chris Clark's blog post about it: http://blog.untrod.com/2016/08/actually-understanding-timezones-in-postgresql.html

textSearchVector :: SqlType Text

textSearchVector defines a type for indexed text searching. It corresponds to the TSVECTOR type in PostgreSQL.

convertSqlType :: (b -> a) -> (a -> b) -> SqlType a -> SqlType b

Migration Guide: convertSqlType retains the same name

convertSqlType changes the Haskell type used by a SqlType in the same manner as maybeConvertSqlType in cases where an a can always be converted to a b.

maybeConvertSqlType :: (b -> a) -> (a -> Maybe b) -> SqlType a -> SqlType b

Migration Guide: maybeConvertSqlType has been replaced with tryConvertSqlType, which allows an error message to be returned when conversion fails.

maybeConvertSqlType changes the Haskell type used by a SqlType without changing the column type that will be used in the database schema. The functions given will be used to convert the now Haskell type to and from the original type when reading and writing values from the database. When reading an a value from the database, the conversion function should produce Nothing if the value cannot be successfully converted to a b.

data TableParams readEntity writeEntity key

MigrationGuide: TableParams no longer exists. See the migration guide for mkTableDefinition

TableParams is the simplest way to make a TableDefinition. You can use mkTableDefinition to make a definition from the simplified params. Where TableDefinition requires the tableFields, tableFromSql, and tableToSql to all be defined separately and kept in sync, TableParams provides a single tblMapper field that specifies all three simultaneously and ensures they are consistent with one another.

Constructors

TableParams

Fields

tblName :: String

The name of the table in the database

tblMapper :: RelationalMap writeEntity readEntity

The relational mapping that defines how the Haskell entity type is converted both to and from sql. The fields utilized in the mapping are used to automatically build the list of FieldDefinitions that define the structure of the table in the database.

tblSafeToDelete :: [String]

A list of any columns that may be deleted from the table by Orville. (Orville will never delete a column without being told it is safe)

tblPrimaryKey :: PrimaryKey key

A function to set the key on the entity

tblGetKey :: readEntity -> key

A function to get the key on the entity

tblComments :: TableComments ()

Any comments that might be interesting for developers to see. These comments will get printed in the log if there is an erro while attempting to migrate the table.

data RelationalMap a b

Migration guide: This type has been replaced with the SqlMarshaller type in the new orville. The interface is similar, though the names of the functions have been updated in many cases. See the migration guides for those functions to find their new names.

fields :: RelationalMap a b -> [SomeField]

Migration Guide: The fields in new orville's SqlMarshaller are somewhat more sophisticated than those of a RelationalMap. The fields function is no longer offered with this simple interface as a result, but the foldMarshallerFields function can be used in combination with the collectFromField helper to collect the desired information from each field.

mapAttr :: (a -> b) -> RelationalMap b c -> RelationalMap a c

Migration Guide: mapAttr has been renamed to marshallNested

mapField :: FieldDefinition nullability a -> RelationalMap a a

Migration Guide: mapField has been removed, though its functional equivalent is marshallReadOnlyField

attrField :: (a -> b) -> FieldDefinition nullability b -> RelationalMap a b

Migration Guide: attrField has been renamed to marshallField

maybeMapper :: RelationalMap a b -> RelationalMap (Maybe a) (Maybe b)

Migration Guide: maybeMapper has been renamed to marshallMaybe

prefixMap :: String -> RelationalMap a b -> RelationalMap a b

Migration Guide: prefixMap has been renamed to prefixMarshaller

partialMap :: RelationalMap a (Either String a) -> RelationalMap a a

Migration Guide: partialMap has been renamed to marshallPartial

readOnlyMap :: RelationalMap a b -> RelationalMap c b

Migration Guide: readOnlyMap has been renamed to marshallReadOnly

readOnlyField :: FieldDefinition nullability a -> RelationalMap b a

Migration Guide: readOnlyField has been renamed to marshallReadOnlyField

data OrvilleEnv conn

Migration Guide: OrvilleEnv has been renamed to OrvilleState. It no longer has any type paremeters. The connection type is fixed and cannot be changed.

OrvilleEnv tracks all the environment information required for an 'OrvilleT conn m' Monad to operate. Use newOrvilleEnv to construct one.

newOrvilleEnv :: Pool conn -> OrvilleEnv conn

Migration Guide: newOrvilleEnv has been renamed to newOrvilleState. The new function requires a parameter to be passed before the connection pool to specify the level of detail to be used when Orville reports errors.

newOrvilleEnv initialized an OrvilleEnv for service. The connection pool provided will be used to obtain connections to the database ase required. You can use the createConnectionPool utility function to create a connection pool to a PosgreSQL server.

setStartTransactionSQL :: String -> OrvilleEnv conn -> OrvilleEnv conn

Migration Guide: setStartTransactionSQL has been renamed to setBeginTransactionExpr

aroundRunningQuery :: (forall a. QueryType -> String -> IO a -> IO a) -> OrvilleEnv conn -> OrvilleEnv conn

Migration Guide: aroundRunningQuery has been renamed to addSqlExecutionCallback

addTransactionCallBack :: (TransactionEvent -> IO ()) -> OrvilleEnv conn -> OrvilleEnv conn

Migration Guide: addTransactionCallBack retains the same name

ormEnvPool :: OrvilleEnv conn -> Pool conn

data TransactionEvent

Constructors

TransactionStart

TransactionCommit

TransactionRollback

data OrvilleT conn m a

Migration Guide: OrvilleT has been removed. In its place you can simply use a ReaderT OrvilleState. If you have another ReaderT layer in your monad stack you can add the OrvilleState to the reader context for that layer instead, which is more efficient than having multiple ReaderT layers. If you have a simple case of OrvilleT conn IO the new Orville offers a simpler Orville monad (not a transformer) to get you started.

unOrvilleT :: OrvilleT conn m a -> ReaderT (OrvilleEnv conn) m a

data SqlValue

class IConnection conn => HasOrvilleContext conn m | m -> conn where

Migration Guide: HasOrvilleContext has been renamed to HasOrvilleState. getOrvilleEnv and localOrvilleEnv have been renamed to askOrvilleState and localOrvilleState.

HasOrvilleContext defines the operations that must be available in your own monad for managing the connection pool that Orville functions will use to access the database and manage transaction state. In most cases you can include OrvilleT in your Monad stack and then automatically derive an instance of HasOrvilleContext.

You could also provide your own implementations of these functions instead of using OrvilleT, if that is the easiest approach for your Monad.

Methods

getOrvilleEnv :: m (OrvilleEnv conn)

getOrvilleEnv fetches the Orville environment from the Monad context. Analogous to ask from the Reader monad.

localOrvilleEnv

Arguments

:: (OrvilleEnv conn -> OrvilleEnv conn)

-> m a

-> m a
localOrvilleEnv locally modifies the Orville environment for the scope of the provided action. This allows Orville to track with a connection is acquired, open transactions, etc. Analogous to local from the Reader monad.

class (Monad m, MonadIO m, HasOrvilleContext conn m, MonadThrow m, MonadOrvilleControl m, MonadFail m) => MonadOrville conn m

Migration Guide: MonadOrville retains the same name, but the conn parameter has been removed. MonadFail and MonadThrow have been removed as superclass constraints.

MonadOrville does not have any methods of its own. Instead it brings all the typeclass constraints required by Orville functions that need to access the database into a single typeclass. In some cases you can include OrvilleT in your Monad stack and then automatically derive an instance of MonadOrville. However, more likely you are using some third party monad somewhere in your stack that does not han a MonadOrvilleControl instance. In this case you won't be able to derive MonadOrville, but providing a simple empty instance will do:

   instance O.MonadOrville Postgres.Connection MyMonad
 

runOrville :: OrvilleT conn m a -> OrvilleEnv conn -> m a

Migration Guide: runOrville now operates on the concrete Orville monad becase OrvilleT has been removed. Assuming you are replacing usages of OrvilleT with ReaderT you will want to replace usages of runOrville with runReaderT.

mapOrvilleT :: Monad n => (m a -> n b) -> OrvilleT conn m a -> OrvilleT conn n b

Migration Guide: mapOrvilleT has been removed because OrvilleT has been removed. If you're replacing OrvilleT with ReaderT then mapOrvilleT should be replaced with mapReaderT.

class MonadOrvilleControl m where

Migration Guide: MonadOrvilleControl retains the same name. The liftFinally member has been removed. There are new liftCatch and liftMask members that must be implemented, however. Instances of the new MonadOrvilleControl are provided for IO and ReaderT. Helper functions for implmenting the members via UnliftIO can be found in Orville.PostgreSQL.UnliftIO.

MonadOrvilleControl provides an interface for the kinds of IO operations that Orville functions need to lift into the Monad providing the MonadOrville instance. This typeclass allows users to provide their own lifting strategies in case the Monad stack in question has special needs. If you are only using ReaderT and OrvilleT layers in your monad stack, you can probably implement this for your own Monad wrapper type using the provided default functions and providing functions to wrap and unwrapper your Monad layer:

   instance MonadOrvilleControl MyMonad where
     liftWithConnection = defaultLiftWithConnection wrapMyMonad unWrapMyMonad
     liftFinally = defaultLiftFinally wrapMyMonad unWrapMyMonad
  

If you are using transformers in your monad stack beyond ReaderT, they probably don't provide MonadOrvilleControl instances (e.g. third party libraries). In this case, see MonadUnliftIO for more help. If you're still stuck (because your library doesn't support MonadTransControl), try MonadBaseControl instead. If you're *still* stuck after that, please file an issue on Github at https://github.com/flipstone/orville so we can can help out!

Methods

liftWithConnection :: (forall a. (conn -> IO a) -> IO a) -> (conn -> m b) -> m b

liftFinally :: (forall a b. IO a -> IO b -> IO a) -> m c -> m d -> m c

defaultLiftWithConnection :: MonadOrvilleControl m => (forall a. m a -> n a) -> (forall a. n a -> m a) -> (forall a. (conn -> IO a) -> IO a) -> (conn -> n b) -> n b

Migration Guide: defaultLiftWithConnection has been removed. In its place you can use either the ReaderT instance of MonadOrvilleControl or the helpers in Orville.PostgreSQL.UnliftIO.

defaultLiftWithConnection provides a simple definition of liftWithConnection for MonadOrvilleControl instances when the Monad in question is a wrapper around a type that already implements MonadOrvilleControl

defaultLiftFinally :: MonadOrvilleControl m => (forall a. m a -> n a) -> (forall a. n a -> m a) -> (forall a b. IO a -> IO b -> IO a) -> n c -> n d -> n c

Migration Guide: defaultLiftWithConnection has been removed (along with liftFinally)

defaultLiftFinally provides a simple definition of liftWithConnection for MonadOrvilleControl instances when the Monad in question is a wrapper around a type that already implements MonadOrvilleControl

data QueryType

Constructors

SelectQuery

InsertQuery

UpdateQuery

DeleteQuery

DDLQuery

withCachedConnection :: MonadOrville conn m => m a -> m a

Migration Guide: withCachedConnection has been renamed to withConnection_

Runs an action with a cached connection. Without using this, or wrapping calls in a transaction using withTransaction, successive calls to functions like insertRecord and updateRecord are *not* guaranteed to occur on the same connection.

withTransaction :: MonadOrville conn m => m a -> m a

Migration Guide: withTransaction retains the same name.

data ColumnFlag

Migration Guide: ColumnFlag has been removed. Depending on flag constructor there may or may not be a replacement.

ColumnDefault - replaced by the setDefaultValue function in new orville Unique - replaced by the addUniqueConstraint function in new orville References - replaced by the addForeignKeyConstraint function in new orville ColumnDescription - removed AssignedByDatabase - removed, though many cases are handled by marshallReadOnlyField

Constructors

ColumnDefault a => Default a

Unique

References (TableDefinition readEntity writeEntity key) (FieldDefinition nullability key)

ColumnDescription String

AssignedByDatabase

class ColumnDefault a where

Methods

toColumnDefaultSql :: a -> String

data Now

Constructors

Now

data FieldDefinition nullability a

Migration Guide: The signature of the FieldDefinition type has not changed, but many of the constructors and accessors have. See the migration guides on individual functions for more info.

data Nullable

Nullable is a values-less type used to track that a FieldDefinition represents a field that is marked nullable in the database schema. See the Nullability type for the value-level representation of field nullability.

data NotNull

'NotNull is a values-less type used to track that a FieldDefinition represents a field that is marked not-null in the database schema. See the Nullability type for the value-level representation of field nullability.

data Nullability nullability where

Nullability represents whether a field will be marked as NULL or 'NOT NULL' in the database schema. It is a GADT so that the value constructors can be used to record this knowledge in the type system as well. This allows functions that work only on Nullable or NotNull fields to indicate this in their type signatures as appropriate.

Constructors

Nullable :: Nullability Nullable

NotNull :: Nullability NotNull

isFieldNullable :: FieldDefinition nullability a -> Bool

Migration Guide: isFieldNullable has been replaced with fieldIsNotNullable, which has the same signture but the Bool returned is the opposite.

fieldOfType :: SqlType a -> String -> FieldDefinition NotNull a

Migration Guide: fieldOfType is essentially unchanged in the new orville.

textField :: String -> Int -> FieldDefinition NotNull Text

Migration Guide: textField has been renamed to boundedTextField. It now takes an Int32 rather than an Int

fixedTextField :: String -> Int -> FieldDefinition NotNull Text

Migration Guide: fixedTextField retains the same name. It now takes an Int32 rather than an Int

unboundedTextField :: String -> FieldDefinition NotNull Text

Migration Guide: unboundedTextField retains the same name.

dayField :: String -> FieldDefinition NotNull Day

Migration Guide: dayField has been renamed to dateField

utcTimeField :: String -> FieldDefinition NotNull UTCTime

Migration Guide: utcTimeField has been renamed to utcTimestampField

int32Field :: String -> FieldDefinition NotNull Int32

Migration guide: int32Field has been renamed to integerField

int64Field :: String -> FieldDefinition NotNull Int64

Migration guide: int64Field has been renamed to bigIntegerField

doubleField :: String -> FieldDefinition NotNull Double

Migration guide: doubleField retains the same name.

boolField :: String -> FieldDefinition NotNull Bool

Migration guide: boolField has been renamed to booleanField

automaticIdField :: String -> FieldDefinition NotNull Int32

Migration guide: automaticIdField has been renamed to serialField

searchVectorField :: String -> FieldDefinition NotNull Text

Migration guide: searchVectorField has been renamed to textSearchVectorField

nullableField :: FieldDefinition NotNull a -> FieldDefinition Nullable (Maybe a)

Migration Guide: nullableField retains the same name

Makes a NotNull field Nullable by wrapping the Haskell type of the field in Maybe. The field will be marked as NULL in the database schema and the value Nothing will be used to represent NULL values when converting to and from sql.

foreignKeyField :: String -> TableDefinition readEntity writeEntity key -> FieldDefinition nullability key -> FieldDefinition nullability key

Migration Guide: foreignKeyField has been removed. It is replaced by addForeignKeyConstraint which adds a foreign key constraint to an existing FieldDefinition.

withFlag :: FieldDefinition nullability a -> ColumnFlag -> FieldDefinition nullability a

Migration Guide: withFlag has been removed. See the migration guide on ColumnFlag regarding the new API.

withName :: FieldDefinition nullability a -> String -> FieldDefinition nullability a

Migration Guide: withName has been removed.

withConversion :: FieldDefinition nullability a -> (SqlType a -> SqlType b) -> FieldDefinition nullability b

Migration Guide: withConversion has been replaced with convertField, whose arguments are flipped from those of withConversion. Note there is also now a coerceField function that can be used with newtype wrappers, provided the constructor is available where coerceField is used.

fieldFromSql :: FieldDefinition nullability a -> FromSql a

Migration Guide: fieldFromSql has been replaced with fieldValueFromSqlValue

fieldToSqlValue :: FieldDefinition nullability a -> a -> SqlValue

Migration Guide: fieldToSqlValue has been renamed to fieldValueToSqlValue

data SomeField

Constructors

SomeField (FieldDefinition nullability a)

withPrefix :: FieldDefinition nullability a -> String -> FieldDefinition nullability a

Migration Guide: withPrefix has been replaced by prefixField whose arguments are flipped relative to withPrefix

fieldName :: FieldDefinition nullability a -> String

fieldType :: FieldDefinition nullability a -> SqlType a

fieldFlags :: FieldDefinition nullability a -> [ColumnFlag]

fieldFlags has been removed. See the new fieldDefaultValue and fieldTableConstraints functions

data IndexDefinition

Constructors

IndexDefinition

Fields

indexName :: String

indexUnique :: Bool

indexTable :: String

indexBody :: String

uniqueIndex :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> IndexDefinition

Migration Guide: uniqueIndex no longer requires a name to be specified. Migration will be done automatically by inspecting the structure of the indexes that exist in the database. It also no longer accepts a TableDefinition at the time of creating the IndexDefinition. Instead you should use addTableIndexes to add the IndexDefinition to the TableDefinition for the table that you wish to index.

If you wish to specify the index name explicitly, you can use uniqueNamedIndex instead. If you do so, index migration will be managed by comparing to the names of existing indexes rather than checking that the index structure matches the Haskell definition.

simpleIndex :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> IndexDefinition

Migration Guide: simpleIndex has been renamed to nonUniqueIndex. It no longer requires a name to be specified. Migration will be done automatically be inspecting the structure of the indexes that exist in the database. It also no longer accepts a TableDefinition at the time of creating the IndexDefinition. Instead you should use addTableIndexes to add the IndexDefinition to the TableDefinition for the table that you wish to index.

If you wish to specify the index name explicitly, you can use nonUniqueNamedIndex instead. If you do so, index migration will be managed by comparing to the names of existing indexes rather than checking that the index structure matches the Haskell definition.

simplePartialIndex :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> [String] -> IndexDefinition

Works much the same as simpleIndex but takes a list of strings that are the conditions of a where clause on index creation for partial indexes

uniquePartialIndex :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> [String] -> IndexDefinition

Works much the same as uniqueIndex but takes a list of strings that are the conditions of a where clause on index creation for partial indexes

data ConstraintDefinition

Constructors

ConstraintDefinition

Fields

constraintName :: String

constraintTable :: String

constraintBody :: String

data SequenceDefinition

Constructors

SequenceDefinition

Fields

sequenceName :: String

sequenceIncrement :: Maybe Int

sequenceMinValue :: Maybe Int

sequenceMaxValue :: Maybe Int

sequenceStart :: Maybe Int

sequenceCache :: Maybe Int

sequenceCycle :: Bool

uniqueConstraint :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> ConstraintDefinition

Migration Guide: uniqueConstraint no longer accepts a name parameter. Instead the constraint is migrated automatically based on the structure of existing constraints found in the database. It also no longer accepts a TableDefinition. Instead you should use addTableConstraints to add the ConstraintDefinition to the table that you wish to apply the constraint to.

dropConstraint :: TableDefinition readEntity writeEntity key -> String -> SchemaItem

Migration Guide: dropConstraint has been removed. Constraints are now dropped automatically during auto-migration when they are removed from the TableDefinition.

data FromSql a

data FromSqlError

Constructors

RowDataError !RowDataErrorDetails
Captures a failure in the translation of a SQL value from a particular field to it's corresponding Haskell values.

MissingColumn !MissingColumnDetails
An expected column was not returned by the database

ConversionError !ConversionErrorDetails
A conversion between haskell representations failed at a point where we don't know what column the value came from. This is the case when using the partialMap combinator.

data RowDataErrorDetails

Constructors

RowDataErrorDetails

Fields

rowErrorReason :: !RowDataErrorReason

rowErrorColumnName :: !String

Column name for the erroneous value

rowErrorPrimaryKeys :: ![(String, SqlValue)]

Primary keys. Empty if not known

data RowDataErrorReason

Constructors

TypeMismatch
Sql value has a different type than expected

Fields

!String

Actual type

!String

Expected type

IntegralOutOfBounds
An integer value was outside the expected bounds.

Fields

!Integer

Lower bound

!Integer

Upper bound

!Integer

Actual value

DecodingFailure !String
Generic decoding failure

data MissingColumnDetails

Constructors

MissingColumnDetails

Fields

missingColumn :: !String

actualColumns :: ![String]

data ConversionErrorDetails

Constructors

ConversionErrorDetails

Fields

convErrorReason :: !String

convErrorPrimaryKeys :: ![(String, SqlValue)]

Primary key value(s). Empty if not known

showFromSqlErrorMinimal :: FromSqlError -> String

Shows the error in a way that should not contain any potentially sensitive data. This is used for the Show instance.

showFromSqlErrorForLogging :: FromSqlError -> String

Shows the error in a way appropriate for logging within an application. The resulting string contains information that is useful for debugging but is potentially undesirable to expose outside of the application (such as primary key values).

showSqlValueType :: SqlValue -> String

User friendly identifier labels for SqlValues

class ColumnSpecifier col where

Methods

selectForm :: col -> SelectForm

col :: (ColumnSpecifier col, Convertible SqlValue a) => col -> FromSql a

data ToSql a b

getField :: Convertible a SqlValue => (entity -> a) -> ToSql entity ()

getComponent :: (entity -> a) -> ToSql a () -> ToSql entity ()

data SchemaItem

Migration Guide: SchemaItem retains the same name. The Index, DropIndex, Constraint and DropConstraint constructors have been removed. These items are now added to the TableDefinition via addTableConstraints and addTableIndexes. The remaining constructors have been prefixed with the word Schema (e.g. Table has been renamed to SchemaTable). There is no explicit replacement for DropIndex and DropConstraint. Orville will automatically drop indexes and constraints that are no longer mentioned on the TableDefinition for any tables that it migrates.

Constructors

Table (TableDefinition readEntity writeEntity key)

DropTable String

Index IndexDefinition

DropIndex String

Constraint ConstraintDefinition

DropConstraint String String

Sequence SequenceDefinition

DropSequence String

type SchemaDefinition = [SchemaItem]

Migration Guide: SchemaDefinition has been removed. Use [SchemaItem] instead.

type Record = Int

Migration Guide: Record has been removed. It's recommended that you create a separate record key type for each of your entities instead.

type CreatedAt = UTCTime

type UpdatedAt = UTCTime

type OccurredAt = UTCTime

data TableComments a

noComments :: TableComments ()

say :: String -> (Int, Int, Int) -> String -> TableComments ()

data WhereCondition

whereAnd :: [WhereCondition] -> WhereCondition

Migration Guide: whereAnd has been removed. Use the binary function andExpr to combine BooleanExpr expressions instead. andExpr is also available as the operator (.&&)

whereOr :: [WhereCondition] -> WhereCondition

Migration Guide: whereOr has been removed. Use the binary function orExpr to combine BooleanExpr expressions instead. orExpr is also available as the operator (.||)

whereIn :: FieldDefinition nullability a -> [a] -> WhereCondition

Migration Guide: whereIn has been renamed to fieldIn. It now takes a NonEmpty list of values to reflect this is a requirement in SQL.

whereLike :: FieldDefinition nullability a -> String -> WhereCondition

Migration Guide: whereLike has been renamed to fieldLike. It now takes a T.Text value rather than a String.

whereLikeInsensitive :: FieldDefinition nullability a -> String -> WhereCondition

Migration Guide: whereLikeInsensitive has been renamed to fieldLikeInsensitive. It now takes a T.Text value rather than a String.

whereNotIn :: FieldDefinition nullability a -> [a] -> WhereCondition

Migration Guide: whereNotIn has been renamed to fieldNotIn. It now takes a NonEmpty list of values to reflect this is a requirement in SQL.

whereQualified :: TableDefinition a b c -> WhereCondition -> WhereCondition

Migration Guide: whereQualified has been removed. If you need qualified column references you can use the SQL building functions found in Orville.PostgreSQL.Expr to build them. The qualifyColumn function can be used to qualify column references in that context. BooleanExpr values built directly this way can be easily used in conjuction with other helpers such as fieldEquals which also build BooleanExpr values themselves.

whereRaw :: String -> [SqlValue] -> WhereCondition

Migration Guide: whereRaw has been removed. In its place you should use the more general functions such as unsafeSqlExpression or unsafeRawSql in the Orville.PostgreSQL.Raw.RawSql module to build a BooleanExpr.

whereToSql :: [WhereCondition] -> (String, [SqlValue])

Migration Guide: whereToSql has been removed. It is replaced by the more general toBytesAndParams function in Orville.PostgreSQL.Raw.RawSql.

isNull :: FieldDefinition Nullable a -> WhereCondition

Migration Guide: isNull has been renamed to fieldIsNull

isNotNull :: FieldDefinition Nullable a -> WhereCondition

Migration Guide: isNotNull has been renamed to fieldIsNotNull

(.==) :: FieldDefinition nullability a -> a -> WhereCondition

(.<>) :: FieldDefinition nullability a -> a -> WhereCondition

(.<-) :: FieldDefinition nullability a -> [a] -> WhereCondition

(%==) :: FieldDefinition nullability a -> a -> WhereCondition

(.>) :: FieldDefinition nullability a -> a -> WhereCondition

(.>=) :: FieldDefinition nullability a -> a -> WhereCondition

(.<) :: FieldDefinition nullability a -> a -> WhereCondition

(.<=) :: FieldDefinition nullability a -> a -> WhereCondition

data SelectOptions

Constructors

SelectOptions

Fields

selectDistinct :: First Bool

selectOptWhere :: [WhereCondition]

selectOptOrder :: [OrderByClause]

selectOptLimit :: First Int

selectOptOffset :: First Int

selectOptGroup :: [GroupByClause]

where_ :: WhereCondition -> SelectOptions

distinct :: SelectOptions

order :: ToOrderBy a => a -> SortDirection -> SelectOptions

limit :: Int -> SelectOptions

offset :: Int -> SelectOptions

groupBy :: ToGroupBy a => a -> SelectOptions

selectOptionsToSql :: SelectOptions -> (String, [SqlValue])

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

data FieldUpdate

fieldUpdate :: FieldDefinition nullability a -> a -> FieldUpdate

(.:=) :: FieldDefinition nullability a -> a -> FieldUpdate

data OrderByClause

Constructors

OrderByClause String [SqlValue] SortDirection

data SortDirection

Constructors

Ascending

Descending

migrateSchema :: MonadOrville conn m => SchemaDefinition -> m ()

Migration Guide: migrateSchema has been renamed to autoMigrateSchema

migrateSchema will attempt to make changes to the actual database schema that it it matches the provided SchemaDefinition. Unsafe migrations such as dropping tables or columns are never attempted unless the SchemaDefinition explicitly states that the items are safe to drop. Column types may be changed, but will fail if the database cannot successfully make the request type change.

data MigrationError

Constructors

MigrationLockExcessiveRetryError String

MigrationExecutionError SchemaItem SomeException

generateMigrationPlan :: MonadOrville conn m => SchemaDefinition -> m (Maybe MigrationPlan)

Migration Guide: generateMigrationPlan retains the same name. It has changed to always return a MigrationPlan. You can use check whether migrationPlanSteps is as empty list if you wish to determine whether any migrations will be performed by the plan.

generateMigrationPlan inspects the state of the actual database schema and constructs a plan describing what changes would be made to make it match the provided SchemaDefinition. If the actual schema already matches the definition, Nothing will be returned.

data MigrationPlan

Migration Guide: MigrationPlan retains the same name.

data MigrationItem

Migration Guide: MigrationItem has been renamed to MigrationStep, which is now a simple RawSql wrapper. You can use RawSql.toExampleBytes if you wish to render it to a bytestring for display purposes.

Constructors

MigrationItem

Fields

migrationItemSchemaItem :: SchemaItem

migrationItemDDL :: DDL

migrationPlanItems :: MigrationPlan -> [MigrationItem]

Migration Guide: migrationPlanItems has been renamed to migrationPlanSteps

data Pagination m entity

Constructors

Pagination

Fields

pageRows :: [entity]

pageNext :: Maybe (m (Pagination m entity))

buildPagination :: (MonadOrville conn m, Bounded orderField, Enum orderField) => TableDefinition readEnt write key -> FieldDefinition NotNull orderField -> (readEnt -> orderField) -> Maybe WhereCondition -> Word -> m (Pagination m readEnt)

selectAll :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> SelectOptions -> m [readEntity]

Migration Guide: selectAll has been renamed to findEntitiesBy

selectFirst :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> SelectOptions -> m (Maybe readEntity)

Migration Guide: selectFirst has been renamed to findFirstEntityBy

deleteRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> key -> m ()

Migration Guide: deleteRecord has been renamed to deleteEntity. Note that there are also new variant functions deleteAndReturnEntity and deleteEntityAndReturnRowCount that return Maybe readEntity and Int respectively.

deleteWhere :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [WhereCondition] -> m Integer

Migration Guide: deleteWhere has been renamed to deleteEntities. It now takes a Maybe BooleanExpr rather than [WhereCondition]

findRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> key -> m (Maybe readEntity)

Migration Guide: findRecord has been renamed to findEntity

findRecords :: (Ord key, MonadOrville conn m) => TableDefinition readEntity writeEntity key -> [key] -> m (Map key readEntity)

Migration Guide: findRecords has been renamed to findEntities. It now requires a NonEmpty key rather than simply [key] and returns a [readEntity] instead of a Map.

findRecordsBy :: (Ord fieldValue, MonadOrville conn m) => TableDefinition readEntity writeEntity key -> FieldDefinition nullability fieldValue -> SelectOptions -> m (Map fieldValue [readEntity])

Migration Guide: findRecordsBy has been renamed to findEntitiesBy. It no longer takes a FieldDefinition to group by. Instead it simply returns a [readEntity]

insertRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> writeEntity -> m readEntity

Migration Guide: insertRecord has been renamed to insertAndReturnEntity. Note there are also new variant functions insertEntity and insertEntityAndReturnRowCount that return () and Int respectively.

insertRecordMany :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [writeEntity] -> m ()

Migration Guide: insertRecordMany has been renamed to insertEntities. It now requires a NonEmpty writeEntity rather than [writeEntity]. Note that there are also new variant functions insertAndReturnEntities and insertEntitiesAndReturnRowCount.

insertRecordManyReturning :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [writeEntity] -> m [readEntity]

Migration Guide: insertRecordManyReturning has been renamed to insertAndReturnEntities.

updateFields :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [FieldUpdate] -> [WhereCondition] -> m Integer

Migration Guide: updateFields has been renamed to updateFieldsAndReturnRowCount, but now takes a NonEmpty SetClause instead of a [Field Update] and a Maybe BooleanExpr instead of a [WhereCondition].

updateFields still exists as a variant of this function, but returns () rather than Int. updateFieldsAndReturnEntities is now available as well.

updateRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> key -> writeEntity -> m ()

Migration Guide: updateRecord has been renamed to updateEntity. Note that there are also new variant functions updateAndReturnEntity and updateEntityAndReturnRowCount@.

sequenceNextVal :: MonadOrville conn m => SequenceDefinition -> m Int

sequenceSetVal :: MonadOrville conn m => SequenceDefinition -> Int -> m Int

sequenceCurrVal :: MonadOrville conn m => SequenceDefinition -> m Int

createIndexesConcurrently :: MonadOrville conn m => [IndexDefinition] -> m ()

Migration Plan: createIndexesConcurrently has been removed. You should now use setIndexCreationStrategy Asynchronous instead.

createIndexesConcurrently will create the given indexes, if they do not exist using the PostgreSQL concurrently feature. However, this does *not* mean the the function happens concurrently. This will wait for PostgreSQL to return, but other operations to the table will be allowed during index creation.

Note: PostgreSQL does not allow CREATE INDEX CONCURRENTLY to appear inside of a transaction. Use this function with care.

dropIndexesConcurrently :: MonadOrville conn m => [String] -> m ()

Migration Guide: dropIndexesConcurrently has been removed.

dropIndexesConcurrently will drop each of the given indexes with the CONCURRENTLY keyword, allowing for other table operations to continue while the index is dropped. However there are several caveats that come with this as noted at https://www.postgresql.org/docs/9.6/sql-dropindex.html . Much like createIndexesConcurrently this cannot be used in a transaction. But further this cannot drop indexes that support UNIQUE or PRIMARY KEY constraints.

Use this with care.