Database.Orville.PostgreSQL.MonadBaseControl

Description

MonadBaseControl provides functions and instances for using MonadOrville and OrvilleT for situations where you need to use MonadBaseControl. If you do not know if you need MonadBaseControl, then you probably don't need to use this module. If you are thinking about using MonadBaseControl instead of MonadUnliftIO, we recommend reading Michael Snoyman's excellent "A Tale of Two Brackets" (https:/www.fpcomplete.comblog201706/tale-of-two-brackets) if you have not already done so.

If you're still here after reading above, this module provides the functions you need to implement MonadOrvilleControl for your Monad stack using its MonadBaseControl instance. The most common way to do this is simply to add the following MonadOrvilleControl instance:

 instance MonadOrvilleControl MyMonad where
   liftWithConnection = liftWithConnectionViaBaseControl
   liftFinally = liftFinallyViaBaseControl

This module also provides a MonadOrvilleControl for StateT as well as MonadBaseControl and MonadTransControl instances for OrvilleT and OrvilleTriggerT.

Documentation

liftWithConnectionViaBaseControl :: MonadBaseControl IO m => (forall b. (conn -> IO b) -> IO b) -> (conn -> m a) -> m a

liftWithConnectionViaBaseControl can be use as the implementation of liftWithConnection for MonadOrvilleControl when the Monad implements MonadBaseControl.

liftFinallyViaBaseControl :: MonadBaseControl IO m => (forall c d. IO c -> IO d -> IO c) -> m a -> m b -> m a

liftFinallyViaBaseControl can be use as the implementation of 'liftFinally for MonadOrvilleControl when the Monad implements MonadBaseControl.

Orphan instances

MonadBaseControl b m => MonadBaseControl b (OrvilleT conn m)
Because we recommend using MonadUnliftIO rather than MonadBaseControl, we do not provide MonadBaseControl instance for OrvilleT by default along with the definition. If you do need to use MonadBaseControl, however, this is the canonical instance for OrvilleT.

Instance details

Associated Types

type StM (OrvilleT conn m) a :: Type

Methods

liftBaseWith :: (RunInBase (OrvilleT conn m) b -> b a) -> OrvilleT conn m a

restoreM :: StM (OrvilleT conn m) a -> OrvilleT conn m a

MonadBaseControl b m => MonadBaseControl b (OrvilleTriggerT trigger conn m)

Instance details

Associated Types

type StM (OrvilleTriggerT trigger conn m) a :: Type

Methods

liftBaseWith :: (RunInBase (OrvilleTriggerT trigger conn m) b -> b a) -> OrvilleTriggerT trigger conn m a

restoreM :: StM (OrvilleTriggerT trigger conn m) a -> OrvilleTriggerT trigger conn m a

MonadTransControl (OrvilleT conn)
Because we recommend using MonadUnliftIO rather than MonadTransControl, we do not provide MonadTransControl instance for OrvilleT by default along with the definition. If you do need to use MonadTransControl, however, this is the canonical instance for OrvilleT.

Instance details

Associated Types

type StT (OrvilleT conn) a :: Type

Methods

liftWith :: Monad m => (Run (OrvilleT conn) -> m a) -> OrvilleT conn m a

restoreT :: Monad m => m (StT (OrvilleT conn) a) -> OrvilleT conn m a

MonadTransControl (OrvilleTriggerT trigger conn)

Instance details

Associated Types

type StT (OrvilleTriggerT trigger conn) a :: Type

Methods

liftWith :: Monad m => (Run (OrvilleTriggerT trigger conn) -> m a) -> OrvilleTriggerT trigger conn m a

restoreT :: Monad m => m (StT (OrvilleTriggerT trigger conn) a) -> OrvilleTriggerT trigger conn m a

MonadBaseControl IO m => MonadOrvilleControl (StateT a m)
Because lifting control operations into StateT is fraught with peril, a MonadOrvilleControl instance for StateT is provided here and implemented via MonadBaseControl rather than together with the MonadOrvilleControl definition. We do not recommend using stateful Monad transformer layers in Monad stacks based on IO. For anyone that must, this is the canonical instance for StateT

Instance details

Methods

liftWithConnection :: (forall a0. (conn -> IO a0) -> IO a0) -> (conn -> StateT a m b) -> StateT a m b

liftFinally :: (forall a0 b. IO a0 -> IO b -> IO a0) -> StateT a m c -> StateT a m d -> StateT a m c