Database.Orville.PostgreSQL.MonadUnliftIO

Description

MonadUnliftIO provides functions and instances for using MonadOrville and OrvilleT for Monad transformer stacks that are using MonadUnliftIO. The most common way to do this is simply to add the following MonadOrvilleControl instance:

 instance MonadOrvilleControl MyMonad where
   liftWithConnection = liftWithConnectionViaUnliftIO
   liftFinally = liftFinallyViaUnliftIO

This module also provides a MonadUnliftIO instance for OrvilleT and OrvilleTrigger. |

Documentation

liftWithConnectionViaUnliftIO :: MonadUnliftIO m => (forall a. (conn -> IO a) -> IO a) -> (conn -> m b) -> m b

liftWithConnectionViaUnliftIO can be use as the implementation of liftWithConnection for MonadOrvilleControl when the Monad implements MonadUnliftIO. |

liftFinallyViaUnliftIO :: MonadUnliftIO m => (forall a b. IO a -> IO b -> IO a) -> m c -> m d -> m c

liftFinallyViaUnliftIO can be use as the implementation of liftFinally for MonadOrvilleControl when the Monad implements MonadUnliftIO. |

Orphan instances

MonadUnliftIO m => MonadUnliftIO (OrvilleT conn m)

Instance details

Methods

askUnliftIO :: OrvilleT conn m (UnliftIO (OrvilleT conn m))

withRunInIO :: ((forall a. OrvilleT conn m a -> IO a) -> IO b) -> OrvilleT conn m b

MonadUnliftIO m => MonadUnliftIO (OrvilleTriggerT trigger conn m)

Instance details

Methods

askUnliftIO :: OrvilleTriggerT trigger conn m (UnliftIO (OrvilleTriggerT trigger conn m))

withRunInIO :: ((forall a. OrvilleTriggerT trigger conn m a -> IO a) -> IO b) -> OrvilleTriggerT trigger conn m b