Database.Orville.PostgreSQL.ResourceT

Description

ResourceT provides ResourceT instance of the Orville typeclasses for situations where you might need it. In particular, if you are using the conduit library, you may want to wrap ResourceT around your normal monad stack, in which case you'll need the MonadOrville instance provided here to use selectConduit.

These instances are not included in the default exports for Orville because the required either a MonadUnliftIO or MonadBaseControl instance of the monad underlying ResourceT, depending on the version of ResourceT you are using. For resource-1.1.10 and above you must provide MonadUnliftIO instance. For versions prior to 1.1.10 you must provide a MonadBaseControl instance.

This is required by MonadOrville requires an instance to MonadBaseControl to be defined. The instance provided here can only use one lifting strategy, one we choose MonadUnliftIO wherever possible (both by our own opinion and because later versions of ResourceT have removed MonadBaseControl support). MonadBaseControl is used for versions of ResourceT before ResourceT supported MonadUnliftIO.

Orphan instances

(MonadUnliftIO m, MonadOrville conn m) => MonadOrville conn (ResourceT m)

Instance details

(Monad m, HasOrvilleContext conn m) => HasOrvilleContext conn (ResourceT m)

Instance details

Methods

getOrvilleEnv :: ResourceT m (OrvilleEnv conn)

localOrvilleEnv :: (OrvilleEnv conn -> OrvilleEnv conn) -> ResourceT m a -> ResourceT m a

(MonadOrvilleControl m, MonadUnliftIO m) => MonadOrvilleControl (ResourceT m)

Instance details

Methods

liftWithConnection :: (forall a. (conn -> IO a) -> IO a) -> (conn -> ResourceT m b) -> ResourceT m b

liftFinally :: (forall a b. IO a -> IO b -> IO a) -> ResourceT m c -> ResourceT m d -> ResourceT m c